+++ lib/lightning_web/live/credential_live/credential_form_component.ex
@@ -9,9 +9,6 @@ defmodule LightningWeb.CredentialLive.CredentialFormComponent do
   alias Lightning.Credentials
   alias LightningWeb.Components.NewInputs
   alias LightningWeb.CredentialLive.GenericOauthComponent
-  alias LightningWeb.CredentialLive.JsonSchemaBodyComponent
-  alias LightningWeb.CredentialLive.OauthComponent
-  alias LightningWeb.CredentialLive.RawBodyComponent
   alias Phoenix.LiveView.JS
 
   @valid_assigns [
@@ -208,6 +205,44 @@ defmodule LightningWeb.CredentialLive.CredentialFormComponent do
     {:noreply, socket |> assign(selected_project: project_id)}
   end
 
+  def handle_event("add_new_project", params, socket) do
+    %{"projectid" => project_id} = params
+
+    project_credentials =
+      fetch_field!(socket.assigns.changeset, :project_credentials)
+
+    project_credentials =
+      if Enum.find(project_credentials, fn pu -> pu.project_id == project_id end) do
+        Enum.map(project_credentials, fn pu ->
+          if pu.project_id == project_id do
+            Ecto.Changeset.change(pu, %{delete: false})
+          else
+            pu
+          end
+        end)
+      else
+        Enum.concat(project_credentials, [
+          %Lightning.Projects.ProjectCredential{project_id: project_id}
+        ])
+      end
+
+    changeset =
+      socket.assigns.changeset
+      |> put_assoc(:project_credentials, project_credentials)
+      |> Map.put(:action, :validate)
+
+    available_projects =
+      filter_available_projects(changeset, socket.assigns.all_projects)
+
+    {:noreply,
+     socket
+     |> assign(
+       changeset: changeset,
+       available_projects: available_projects,
+       selected_project: nil
+     )}
+  end
+
   def handle_event("delete_project", %{"projectid" => project_id}, socket) do
     project_credentials =
       fetch_field!(socket.assigns.changeset, :project_credentials)
@@ -366,6 +401,7 @@ defmodule LightningWeb.CredentialLive.CredentialFormComponent do
           projects={@all_projects}
           users={@users}
           allow_credential_transfer={@allow_credential_transfer}
+          parent_component={@myself}
           return_to={@return_to}
         />
       </.modal>
@@ -400,7 +436,7 @@ defmodule LightningWeb.CredentialLive.CredentialFormComponent do
           phx-change="validate"
           phx-submit="save"
         >
-          <.form_component
+          <LightningWeb.Components.Credentials.form_component
             :let={{fieldset, _valid?}}
             id={@credential.id || "new"}
             schema={@schema}
@@ -478,7 +514,7 @@ defmodule LightningWeb.CredentialLive.CredentialFormComponent do
                     Control which projects have access to this credentials
                   </p>
                   <div class="mt-4">
-                    <.project_credentials
+                    <LightningWeb.Components.Credentials.project_credentials
                       form={f}
                       projects={@all_projects}
                       selected={@selected_project}
@@ -491,10 +527,13 @@ defmodule LightningWeb.CredentialLive.CredentialFormComponent do
                 :if={@action == :edit and @allow_credential_transfer}
                 class="space-y-4"
               >
-                <.credential_transfer form={f} users={@users} />
+                <LightningWeb.Components.Credentials.credential_transfer
+                  form={f}
+                  users={@users}
+                />
               </div>
             </div>
-          </.form_component>
+          </LightningWeb.Components.Credentials.form_component>
           <.modal_footer class="mt-6 mx-6">
             <div class="sm:flex sm:flex-row-reverse">
               <button
@@ -519,166 +558,6 @@ defmodule LightningWeb.CredentialLive.CredentialFormComponent do
         </.form>
       </.modal>
     </div>
-    """
-  end
-
-  attr :id, :string, required: false
-  attr :type, :string, required: true
-  attr :form, :map, required: true
-  attr :action, :any, required: false
-  attr :phx_target, :any, default: nil
-  attr :schema, :string, required: false
-  attr :sandbox_value, :boolean, default: false
-  attr :api_version, :string, default: ""
-  attr :update_body, :any, required: false
-  attr :scopes_changed, :boolean, required: false
-  attr :oauth_clients, :list, required: false
-  slot :inner_block
-
-  defp form_component(%{type: "googlesheets"} = assigns) do
-    ~H"""
-    <OauthComponent.fieldset
-      :let={l}
-      id={@id}
-      form={@form}
-      action={@action}
-      schema={@schema}
-      update_body={@update_body}
-    >
-      <%= render_slot(@inner_block, l) %>
-    </OauthComponent.fieldset>
-    """
-  end
-
-  defp form_component(%{type: "salesforce_oauth"} = assigns) do
-    ~H"""
-    <OauthComponent.fieldset
-      :let={l}
-      id={@id}
-      form={@form}
-      action={@action}
-      schema={@schema}
-      update_body={@update_body}
-      sandbox_value={@sandbox_value}
-      api_version={@api_version}
-      scopes_changed={@scopes_changed}
-    >
-      <%= render_slot(@inner_block, l) %>
-    </OauthComponent.fieldset>
-    """
-  end
-
-  defp form_component(%{type: "raw"} = assigns) do
-    ~H"""
-    <RawBodyComponent.fieldset :let={l} form={@form}>
-      <%= render_slot(@inner_block, l) %>
-    </RawBodyComponent.fieldset>
-    """
-  end
-
-  defp form_component(%{type: _schema} = assigns) do
-    ~H"""
-    <JsonSchemaBodyComponent.fieldset :let={l} form={@form}>
-      <%= render_slot(@inner_block, l) %>
-    </JsonSchemaBodyComponent.fieldset>
-    """
-  end
-
-  attr :projects, :list, required: true
-  attr :selected, :map, required: true
-  attr :phx_target, :any, default: nil
-  attr :form, :map, required: true
-
-  defp project_credentials(assigns) do
-    ~H"""
-    <div class="col-span-3">
-      <%= Phoenix.HTML.Form.label(@form, :project_credentials, "Project Access",
-        class: "block text-sm font-medium text-secondary-700"
-      ) %>
-
-      <div class="flex w-full items-center gap-2 pb-3 mt-1">
-        <div class="grow">
-          <LightningWeb.Components.Form.select_field
-            form={:selected_project}
-            name={:id}
-            values={@projects}
-            value={@selected}
-            prompt=""
-            phx-change="select_item"
-            phx-target={@phx_target}
-            id={"project_credentials_list_for_#{@form[:id].value}"}
-          />
-        </div>
-        <div class="grow-0 items-right">
-          <.button
-            id={"add-new-project-button-to-credential-#{@form[:id].value}"}
-            disabled={@selected == ""}
-            phx-target={@phx_target}
-            phx-value-projectid={@selected}
-            phx-click="add_new_project"
-          >
-            Add
-          </.button>
-        </div>
-      </div>
-
-      <.inputs_for :let={project_credential} field={@form[:project_credentials]}>
-        <%= if project_credential[:delete].value != true do %>
-          <div class="flex w-full gap-2 items-center pb-2">
-            <div class="grow">
-              <%= project_name(@projects, project_credential[:project_id].value) %>
-              <.old_error field={project_credential[:project_id]} />
-            </div>
-            <div class="grow-0 items-right">
-              <.button
-                id={"delete-project-credential-#{@form[:id].value}-button"}
-                phx-target={@phx_target}
-                phx-value-projectid={project_credential[:project_id].value}
-                phx-click="delete_project"
-              >
-                Remove
-              </.button>
-            </div>
-          </div>
-        <% end %>
-        <.input type="hidden" field={project_credential[:project_id]} />
-        <.input
-          type="hidden"
-          field={project_credential[:delete]}
-          value={to_string(project_credential[:delete].value)}
-        />
-      </.inputs_for>
-    </div>
-    """
-  end
-
-  attr :users, :list, required: true
-  attr :form, :map, required: true
-
-  defp credential_transfer(assigns) do
-    ~H"""
-    <div class="hidden sm:block" aria-hidden="true">
-      <div class="border-t border-secondary-200 mb-6"></div>
-    </div>
-    <fieldset>
-      <legend class="contents text-base font-medium text-gray-900">
-        Transfer Ownership
-      </legend>
-      <p class="text-sm text-gray-500">
-        Assign ownership of this credential to someone else.
-      </p>
-      <div class="mt-4">
-        <%= Phoenix.HTML.Form.label(@form, :owner,
-          class: "block text-sm font-medium text-secondary-700"
-        ) %>
-        <LightningWeb.Components.Form.select_field
-          form={@form}
-          name={:user_id}
-          values={@users}
-        />
-        <.old_error field={@form[:user_id]} />
-      </div>
-    </fieldset>
     """
   end
 
@@ -784,12 +663,6 @@ defmodule LightningWeb.CredentialLive.CredentialFormComponent do
   # requires a `pid`
   defp update_body(pid, id, body) do
     send_update(pid, __MODULE__, id: id, body: body)
-  end
-
-  defp project_name(projects, id) do
-    Enum.find_value(projects, fn {name, project_id} ->
-      if project_id == id, do: name
-    end)
   end
 
   defp save_credential(
